split out a separate chunking unit test
